Castor3D: Moved LightSurface to its own file.
