+++ source/Core/Castor3D/Shader/Shaders/GlslShadow.cpp
@@ -5,6 +5,7 @@
 #include "Castor3D/Render/ShadowMap/ShadowMapPassDirectional.hpp"
 #include "Castor3D/Render/ShadowMap/ShadowMapPassSpot.hpp"
 #include "Castor3D/Shader/Shaders/GlslLight.hpp"
+#include "Castor3D/Shader/Shaders/GlslLightSurface.hpp"
 #include "Castor3D/Shader/Shaders/GlslOutputComponents.hpp"
 #include "Castor3D/Shader/Shaders/GlslRay.hpp"
 #include "Castor3D/Shader/Shaders/GlslSurface.hpp"
