+++ source/Plugins/Generic/AtmosphereScattering/AtmosphereLightingModel.cpp
@@ -5,6 +5,7 @@
 
 #include <Castor3D/Shader/Shaders/GlslBRDFHelpers.hpp>
 #include <Castor3D/Shader/Shaders/GlslLight.hpp>
+#include <Castor3D/Shader/Shaders/GlslLightSurface.hpp>
 #include <Castor3D/Shader/Shaders/GlslMaterial.hpp>
 #include <Castor3D/Shader/Shaders/GlslOutputComponents.hpp>
 #include <Castor3D/Shader/Shaders/GlslReflection.hpp>
